+++ src/main/java/pl/thedeem/intellij/common/components/table/reordering/TableColumnItem.java
@@ -1,0 +1,4 @@
+package pl.thedeem.intellij.common.components.table.reordering;
+
+record TableColumnItem(String name, boolean visible) {
+}