+++ src/main/java/pl/thedeem/intellij/dql/fileProviders/DQLResultVirtualFile.java
@@ -1,6 +1,10 @@
 package pl.thedeem.intellij.dql.fileProviders;
 
+import com.intellij.openapi.actionSystem.ActionManager;
+import com.intellij.openapi.actionSystem.ActionToolbar;
 import com.intellij.openapi.project.Project;
+import com.intellij.util.ui.JBUI;
+import com.intellij.util.ui.components.BorderLayoutPanel;
 import org.jetbrains.annotations.NotNull;
 import pl.thedeem.intellij.common.sdk.model.DQLPollResponse;
 import pl.thedeem.intellij.dql.exec.panel.DQLExecutionResult;
@@ -14,8 +18,14 @@ public class DQLResultVirtualFile extends DQLVirtualFile<DQLPollResponse> {
 
     @Override
     public @NotNull JComponent createComponent(@NotNull Project project) {
-        DQLExecutionResult panel = new DQLExecutionResult(project);
-        panel.update(content);
+        BorderLayoutPanel panel = new BorderLayoutPanel();
+        panel.setBorder(JBUI.Borders.empty());
+        DQLExecutionResult result = new DQLExecutionResult(project);
+        result.update(content);
+        ActionToolbar toolbar = ActionManager.getInstance().createActionToolbar("DQL.ResultToolbar", result.getToolbarActions(), true);
+        toolbar.setTargetComponent(panel);
+        panel.addToTop(toolbar.getComponent());
+        panel.addToCenter(result);
         return panel;
     }
 }