+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Small UI improvements:
   - DQL fragment selected during the query execution will now have a green border instead of a selection background
+  - Adding data toolbar for DQL execution results opened as a new editor tab
+  - Adding an option to change the DQL execution results column order and visibility
+- The "Show DQL query" option when executing DQL will now try to show the parsed query instead of the raw one, if
+  possible.
 - Adding support for "Expression DQL" file (with `.edql` extensions), which allows to define a DQL expression without
   the command context.
   Such expression are commonly used in as-a-code DQL queries or in OpenPipeline configurations.